Move rft observation and responses tests to its own file
