+++ tests/ert/unit_tests/storage/test_local_storage.py
@@ -3,7 +3,7 @@ import logging
 import os
 import shutil
 import stat
-from datetime import date, datetime, timedelta
+from datetime import datetime, timedelta
 from pathlib import Path
 from unittest.mock import MagicMock, PropertyMock, patch
 
@@ -20,7 +20,6 @@ from ert.config import (
 from ert.config._observations import (
     BreakthroughObservation,
 )
-from ert.config.rft_config import RFTConfig
 from ert.dark_storage.common import ErtStoragePermissionError
 from ert.storage import (
     ErtStorageException,
@@ -31,7 +30,6 @@ from ert.storage import (
 from ert.storage.local_storage import _LOCAL_STORAGE_VERSION, LocalStorage
 from ert.storage.mode import ModeError
 from tests.ert.defaults_generator import (
-    _create_rft_observation,
     _create_summary_observation,
 )
 from tests.ert.unit_tests.storage._storage_test_helpers import (
@@ -676,349 +674,6 @@ def test_that_breakthrough_observations_and_responses_are_joined_in_endpoint(tmp
         assert obs_and_responses["0"].to_list() == [-1.5]
 
 
-def rft_response(
-    *,
-    well: tuple[str, ...] = ("WELL",),
-    date: tuple[date, ...] = (datetime(2000, 1, 1).date(),),  # noqa: DTZ001
-    prop: tuple[str, ...] = ("SWAT",),
-    depth: tuple[float, ...] = (1006.6,),
-    values: tuple[float, ...] = (100.0,),
-    well_connection_cell: tuple[tuple[int, int, int], ...] = ((10, 10, 10),),
-    cell_center: tuple[tuple[float, float, float], ...] = ((100.0, 105.0, 1000.0),),
-    cell_zones: tuple[tuple[str, ...], ...] = (("zone1",),),
-) -> pl.DataFrame:
-    return (
-        pl.DataFrame(
-            {
-                "well": well,
-                "property": prop,
-                "time": date,
-                "depth": pl.Series(depth, dtype=pl.Float32),
-                "values": pl.Series(values, dtype=pl.Float32),
-                "well_connection_cell": pl.Series(
-                    well_connection_cell, dtype=pl.Array(pl.Int64, 3)
-                ),
-                "cell_center": pl.Series(cell_center, dtype=pl.Array(pl.Float32, 3)),
-                "cell_zones": pl.Series(cell_zones, dtype=pl.List(pl.String)),
-            }
-        )
-        .with_columns(pl.col("time").dt.to_string("%Y-%m-%d").alias("date"))
-        .with_columns(
-            pl.concat_str(
-                [pl.col("well"), pl.col("date"), pl.col("property")], separator=":"
-            ).alias("response_key"),
-        )
-        .select(RFTConfig.response_schema().keys())
-        .pipe(RFTConfig._assert_schema, RFTConfig.response_schema())
-    )
-
-
-def rft_observation(
-    *,
-    name="RFT_OBS1",
-    well="WELL",
-    date="2000-01-01",
-    prop="PRESSURE",
-    value=100.0,
-    error=10.0,
-    east=100.0,
-    north=105.0,
-    tvd=1000.0,
-    zone=None,
-):
-    return _create_rft_observation(
-        name=name,
-        well=well,
-        date=date,
-        prop=prop,
-        value=value,
-        error=error,
-        east=east,
-        north=north,
-        tvd=tvd,
-        zone=zone,
-    )
-
-
-def rft_observation1(*, zone=None):
-    return rft_observation(
-        prop="SWAT",
-        zone=zone,
-    )
-
-
-def rft_observation2(*, well="WELL", zone=None):
-    return rft_observation(
-        name="RFT_OBS2",
-        well=well,
-        prop="SWAT",
-        east=300.0,
-        north=405.0,
-        tvd=2000.0,
-        zone=zone,
-    )
-
-
-def location_metadata(
-    *,
-    east=(100.0, 300.0),
-    north=(105.0, 405.0),
-    tvd=(1000.0, 2000.0),
-    actual_zones=(("zone1",), ("zone2",)),
-    well_connection_cell=((10, 10, 10), (10, 10, 10)),
-) -> pl.DataFrame:
-    df = pl.DataFrame(
-        {
-            "east": east,
-            "north": north,
-            "tvd": tvd,
-            "actual_zones": actual_zones,
-            "well_connection_cell": well_connection_cell,
-            "well_connection_cell_center": pl.Series(
-                [(e, n, t) for e, n, t in zip(east, north, tvd, strict=True)],
-                dtype=pl.Array(pl.Float32, 3),
-            ),
-        },
-        schema=RFTConfig.location_metadata_schema(),
-    )
-    RFTConfig._assert_schema(df, RFTConfig.location_metadata_schema())
-    return df
-
-
-def test_that_get_observations_and_responses_applies_rft_metadata(tmp_path):
-    with open_storage(tmp_path, mode="w") as storage:
-        rft_config = RFTConfig(
-            input_files=["BASE.RFT"],
-            data_to_read={"WELL": {"2000-01-01": ["SWAT"]}},
-        )
-
-        obs1 = rft_observation1()
-        obs2 = rft_observation2()
-
-        experiment = storage.create_experiment(
-            experiment_config={
-                "response_configuration": [rft_config.model_dump(mode="json")],
-                "observations": [
-                    obs1.model_dump(mode="json"),
-                    obs2.model_dump(mode="json"),
-                ],
-            }
-        )
-
-        ensemble = storage.create_ensemble(
-            experiment, ensemble_size=2, iteration=0, name="prior"
-        )
-
-        ensemble.save_response("rft", rft_response(values=(200.0,)), 0)
-        ensemble.save_observation_location_metadata(location_metadata(), 0)
-
-        ensemble.save_response("rft", rft_response(values=(300.0,)), 1)
-        ensemble.save_observation_location_metadata(location_metadata(), 1)
-
-        iens_active_index = np.array([0, 1])
-        active_observations = ["RFT_OBS1"]
-
-        obs_and_responses = ensemble.get_observations_and_responses(
-            active_observations, iens_active_index
-        )
-
-        assert obs_and_responses["response_key"].to_list() == ["WELL:2000-01-01:SWAT"]
-        assert obs_and_responses["index"].to_list() == ["100.0, 105.0, 1000.0, None"]
-        assert obs_and_responses["observation_key"].to_list() == active_observations
-        assert obs_and_responses["0"].to_list() == [200.0]
-        assert obs_and_responses["1"].to_list() == [300.0]
-
-
-def test_that_get_observations_and_responses_disables_observation(tmp_path):
-    with open_storage(tmp_path, mode="w") as storage:
-        rft_config = RFTConfig(
-            input_files=["BASE.RFT"],
-            data_to_read={"WELL": {"2000-01-01": ["SWAT"]}},
-        )
-
-        obs1 = rft_observation1(zone="zone2")
-        obs2 = rft_observation2(zone="zone3")
-
-        experiment = storage.create_experiment(
-            experiment_config={
-                "response_configuration": [rft_config.model_dump(mode="json")],
-                "observations": [
-                    obs1.model_dump(mode="json"),
-                    obs2.model_dump(mode="json"),
-                ],
-            }
-        )
-
-        ensemble = storage.create_ensemble(
-            experiment, ensemble_size=2, iteration=0, name="prior"
-        )
-
-        ensemble.save_response(
-            "rft", rft_response(values=(200.0,), cell_zones=(("zone1",),)), 0
-        )
-        ensemble.save_observation_location_metadata(
-            location_metadata(actual_zones=(("zone1",), ("zone3",))), 0
-        )
-
-        ensemble.save_response(
-            "rft", rft_response(values=(300.0,), cell_zones=(("zone2",),)), 1
-        )
-        ensemble.save_observation_location_metadata(
-            location_metadata(actual_zones=(("zone2",), ("zone3",))), 1
-        )
-
-        iens_active_index = np.array([0, 1])
-        active_observations = ["RFT_OBS1", "RFT_OBS2"]
-
-        obs_and_responses = ensemble.get_observations_and_responses(
-            active_observations, iens_active_index
-        )
-
-        assert obs_and_responses["observation_key"].to_list() == active_observations
-        msg = "expected zone 'zone2' did not match any of the simulated zones: zone1"
-        assert obs_and_responses["0"].to_list() == [None, 200.0]
-        assert obs_and_responses["1"].to_list() == [300.0, 300.0]
-        assert obs_and_responses["qc_error_0"].to_list() == [msg, None]
-        assert obs_and_responses["qc_error_1"].to_list() == [None, None]
-
-
-def test_that_get_observations_and_responses_combines_error_messages(tmp_path):
-    with open_storage(tmp_path, mode="w") as storage:
-        rft_config = RFTConfig(
-            input_files=["BASE.RFT"],
-            data_to_read={"WELL": {"2000-01-01": ["SWAT"]}},
-        )
-
-        obs1 = rft_observation1(zone="zone1")
-        obs2 = rft_observation2(zone="zone2")
-
-        experiment = storage.create_experiment(
-            experiment_config={
-                "response_configuration": [rft_config.model_dump(mode="json")],
-                "observations": [
-                    obs1.model_dump(mode="json"),
-                    obs2.model_dump(mode="json"),
-                ],
-            }
-        )
-
-        ensemble = storage.create_ensemble(
-            experiment, ensemble_size=2, iteration=0, name="prior"
-        )
-
-        ensemble.save_response(
-            "rft", rft_response(values=(200.0,), cell_zones=(("zone1",),)), 0
-        )
-        ensemble.save_observation_location_metadata(
-            location_metadata(
-                actual_zones=(("zone0",), ("zone2",)),
-                well_connection_cell=(None, (10, 10, 10)),
-            ),
-            0,
-        )
-
-        ensemble.save_response(
-            "rft", rft_response(values=(300.0,), cell_zones=(("zone2",),)), 1
-        )
-        ensemble.save_observation_location_metadata(
-            location_metadata(
-                actual_zones=(("zone1",), ("zone2",)),
-                well_connection_cell=(None, (10, 10, 10)),
-            ),
-            1,
-        )
-
-        iens_active_index = np.array([0, 1])
-        active_observations = ["RFT_OBS1", "RFT_OBS2"]
-
-        obs_and_responses = ensemble.get_observations_and_responses(
-            active_observations, iens_active_index
-        )
-
-        assert obs_and_responses["observation_key"].to_list() == active_observations
-
-        msg1 = "expected zone 'zone1' did not match any of the simulated zones: zone0"
-        msg2 = "did not find grid coordinate for location 100.0, 105.0, 1000.0"
-        msg3 = (
-            "no response matched observation data: "
-            "response_key=WELL:2000-01-01:SWAT, well_connection_cell=None"
-        )
-        msg_real0 = f"{msg1};\n{msg2};\n{msg3}"
-        msg_real1 = f"{msg2};\n{msg3}"
-        assert obs_and_responses["0"].to_list() == [None, 200.0]
-        assert obs_and_responses["1"].to_list() == [None, 300.0]
-        assert obs_and_responses["qc_error_0"].to_list() == [msg_real0, None]
-        assert obs_and_responses["qc_error_1"].to_list() == [msg_real1, None]
-
-
-@pytest.mark.parametrize(
-    ("obs2_kwargs", "response_kwargs", "meta_kwargs"),
-    [
-        pytest.param(
-            {"well": "OTHER_WELL"},
-            {"well": ("OTHER_WELL",)},
-            {},
-            id="on response key",
-        ),
-        pytest.param(
-            {},
-            {"well_connection_cell": ([11, 11, 11],)},
-            {"well_connection_cell": ([10, 10, 10], [11, 11, 11])},
-            id="on match key",
-        ),
-    ],
-)
-def test_that_get_observations_and_responses_adds_qc_error_on_rft_mismatch(
-    tmp_path, obs2_kwargs, response_kwargs, meta_kwargs
-):
-    with open_storage(tmp_path, mode="w") as storage:
-        rft_config = RFTConfig(
-            input_files=["BASE.RFT"],
-            data_to_read={"WELL": {"2000-01-01": ["SWAT"]}},
-        )
-
-        obs1 = rft_observation1()
-        obs2 = rft_observation2(**obs2_kwargs)
-
-        experiment = storage.create_experiment(
-            experiment_config={
-                "response_configuration": [rft_config.model_dump(mode="json")],
-                "observations": [
-                    obs1.model_dump(mode="json"),
-                    obs2.model_dump(mode="json"),
-                ],
-            }
-        )
-
-        ensemble = storage.create_ensemble(
-            experiment, ensemble_size=2, iteration=0, name="prior"
-        )
-
-        ensemble.save_response("rft", rft_response(**response_kwargs), 0)
-        ensemble.save_response("rft", rft_response(**response_kwargs), 1)
-
-        ensemble.save_observation_location_metadata(location_metadata(**meta_kwargs), 0)
-        ensemble.save_observation_location_metadata(location_metadata(**meta_kwargs), 1)
-
-        iens_active_index = np.array([0, 1])
-        active_observations = ["RFT_OBS1", "RFT_OBS2"]
-
-        obs_and_responses = ensemble.get_observations_and_responses(
-            active_observations, iens_active_index
-        )
-
-        assert obs_and_responses["observation_key"].to_list() == active_observations
-
-        msg = (
-            "no response matched observation data: "
-            "response_key=WELL:2000-01-01:SWAT, well_connection_cell=[10, 10, 10]"
-        )
-        np.testing.assert_equal(obs_and_responses["0"].to_list(), [None, 100.0])
-        np.testing.assert_equal(obs_and_responses["1"].to_list(), [None, 100.0])
-        assert obs_and_responses["qc_error_0"].to_list() == [msg, None]
-        assert obs_and_responses["qc_error_1"].to_list() == [msg, None]
-
-
 def summary_response(*, key="FOPR", time="2000-01-01", values=100.0) -> pl.DataFrame:
     return pl.DataFrame(
         {
@@ -1091,93 +746,3 @@ def test_that_get_observations_and_responses_adds_qc_error_on_summary_mismatch(
         np.testing.assert_equal(obs_and_responses["1"].to_list(), [None])
         assert obs_and_responses["qc_error_0"].to_list() == [msg]
         assert obs_and_responses["qc_error_1"].to_list() == [msg]
-
-
-@pytest.mark.parametrize(
-    (
-        "approximate_missing_values",
-        "missing_required_columns",
-        "expected_response_values",
-    ),
-    [
-        pytest.param(True, False, [110.0, 310.0], id="interpolate enabled"),
-        pytest.param(False, False, [None, None], id="interpolate disabled"),
-        pytest.param(True, True, [None, None], id="missing required columns"),
-    ],
-)
-def test_that_get_observations_and_responses_interpolates_rft_values(
-    tmp_path,
-    approximate_missing_values,
-    missing_required_columns,
-    expected_response_values,
-):
-    with open_storage(tmp_path, mode="w") as storage:
-        rft_config = RFTConfig(
-            input_files=["BASE.RFT"],
-            data_to_read={"WELL": {"2000-01-01": ["PRESSURE"]}},
-            approximate_missing_values=approximate_missing_values,
-        )
-
-        obs1 = rft_observation(name="RFT_OBS1", value=100.0, tvd=1000.0, zone="zone1")
-        obs2 = rft_observation(name="RFT_OBS2", value=200.0, tvd=2000.0, zone="zone1")
-
-        experiment = storage.create_experiment(
-            experiment_config={
-                "response_configuration": [rft_config.model_dump(mode="json")],
-                "observations": [
-                    obs1.model_dump(mode="json"),
-                    obs2.model_dump(mode="json"),
-                ],
-            }
-        )
-
-        ensemble = storage.create_ensemble(
-            experiment, ensemble_size=1, iteration=0, name="prior"
-        )
-
-        date = datetime(2000, 1, 1).date()  # noqa: DTZ001
-
-        # If the response is missing the required columns for interpolation,
-        # we should skip interpolation and return None, even if interpolation is
-        # enabled. The test drops the columns to simulate a legacy response without
-        # these columns.
-        drop_columns = ["cell_center", "cell_zones"] if missing_required_columns else []
-        ensemble.save_response(
-            "rft",
-            rft_response(
-                well=("WELL", "WELL"),
-                date=(date, date),
-                prop=("PRESSURE", "PRESSURE"),
-                depth=(500.0, 1500.0),
-                values=(50.0, 150.0),
-                well_connection_cell=((10, 10, 10), (10, 10, 12)),
-                cell_center=((100.0, 105.0, 700.0), (100.0, 105.0, 1200.0)),
-                cell_zones=(("zone1",), ("zone1",)),
-            ).drop(drop_columns),
-            0,
-        )
-        ensemble.save_observation_location_metadata(
-            location_metadata(
-                east=(100.0, 100.0),
-                north=(105.0, 105.0),
-                tvd=(1000.0, 2000.0),
-                actual_zones=(("zone1",), ("zone1",)),
-                well_connection_cell=((10, 10, 11), (10, 10, 13)),
-            ),
-            0,
-        )
-
-        iens_active_index = np.array([0])
-        active_observations = ["RFT_OBS1", "RFT_OBS2"]
-
-        obs_and_responses = ensemble.get_observations_and_responses(
-            active_observations, iens_active_index
-        )
-
-        assert obs_and_responses["response_key"].to_list() == [
-            "WELL:2000-01-01:PRESSURE",
-            "WELL:2000-01-01:PRESSURE",
-        ]
-        np.testing.assert_array_equal(
-            obs_and_responses["0"].to_list(), expected_response_values
-        )
